[DeadCode] Rename RemoveUnusedClassConstantRector to RemoveUnusedPrivateClassConstantRector
