--- rules-tests/DeadCode/Rector/ClassConst/RemoveUnusedClassConstantRector/Fixture/external_public.php.inc
@@ -1,20 +1,0 @@
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\ClassConst\RemoveUnusedClassConstantRector\Fixture;
-
-class ExternalPublic
-{
-    public const SOME_CONST = 'dead';
-}
-
-?>
------
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\ClassConst\RemoveUnusedClassConstantRector\Fixture;
-
-class ExternalPublic
-{
-}
-
-?>
