--- rules-tests/DeadCode/Rector/ClassConst/RemoveUnusedClassConstantRector/Fixture/fixture.php.inc
@@ -1,27 +1,0 @@
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\ClassConst\RemoveUnusedClassConstantRector\Fixture;
-
-class Fixture
-{
-    private const SOME_CONST = 'dead';
-
-    public function run()
-    {
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\ClassConst\RemoveUnusedClassConstantRector\Fixture;
-
-class Fixture
-{
-    public function run()
-    {
-    }
-}
-
-?>
