--- rules-tests/DeadCode/Rector/ClassConst/RemoveUnusedClassConstantRector/Fixture/skip_multiple_constants.php.inc
@@ -1,8 +1,0 @@
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\ClassConst\RemoveUnusedClassConstantRector\Fixture;
-
-class SkipMultipleConstants
-{
-    public const SOME_CONST = 'dead', NOT_USED_EITHER = 'dead';
-}
