--- rules-tests/DeadCode/Rector/ClassConst/RemoveUnusedClassConstantRector/Fixture/skip_on_class_api.php.inc
@@ -1,11 +1,0 @@
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\ClassConst\RemoveUnusedClassConstantRector\Fixture;
-
-/**
- * @api
- */
-class SkipOnClassApi
-{
-    public const USED_PUBLICLY_IN_ANOTHER_PROJECT = 'publicly';
-}
