--- rules-tests/DeadCode/Rector/ClassConst/RemoveUnusedClassConstantRector/FixtureOpenSource/skip_public.php.inc
@@ -1,8 +1,0 @@
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\ClassConst\RemoveUnusedClassConstantRector\Fixture;
-
-class SkipPublic
-{
-    public const SOME_CONST = 'dead';
-}
