--- rules-tests/DeadCode/Rector/ClassConst/RemoveUnusedClassConstantRector/Source/AbstractEnum.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\Tests\DeadCode\Rector\ClassConst\RemoveUnusedClassConstantRector\Source;
-
-abstract class AbstractEnum
-{
-
-}
