+++ rules-tests/DeadCode/Rector/ClassConst/RemoveUnusedPrivateClassConstantRector/Fixture/skip_multiple_constants.php.inc
@@ -1,0 +1,8 @@
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\ClassConst\RemoveUnusedPrivateClassConstantRector\Fixture;
+
+class SkipMultipleConstants
+{
+    private const SOME_CONST = 'dead', NOT_USED_EITHER = 'dead';
+}
