+++ rules-tests/DeadCode/Rector/ClassConst/RemoveUnusedPrivateClassConstantRector/Fixture/skip_on_class_api.php.inc
@@ -1,0 +1,11 @@
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\ClassConst\RemoveUnusedPrivateClassConstantRector\Fixture;
+
+/**
+ * @api
+ */
+class SkipOnClassApi
+{
+    private const USED_PUBLICLY_IN_ANOTHER_PROJECT = 'publicly';
+}
