+++ rules-tests/DeadCode/Rector/ClassConst/RemoveUnusedPrivateClassConstantRector/FixtureOpenSource/skip_public.php.inc
@@ -1,0 +1,8 @@
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\ClassConst\RemoveUnusedPrivateClassConstantRector\Fixture;
+
+class SkipPublic
+{
+    public const SOME_CONST = 'dead';
+}
