+++ rules-tests/DeadCode/Rector/ClassConst/RemoveUnusedPrivateClassConstantRector/Source/AbstractEnum.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Tests\DeadCode\Rector\ClassConst\RemoveUnusedPrivateClassConstantRector\Source;
+
+abstract class AbstractEnum
+{
+
+}
