+++ rules-tests/DeadCode/Rector/ClassConst/RemoveUnusedPrivateClassConstantRector/config/project_open_source.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 use Rector\Core\Configuration\Option;
 use Rector\Core\ValueObject\ProjectType;
-use Rector\DeadCode\Rector\ClassConst\RemoveUnusedClassConstantRector;
+use Rector\DeadCode\Rector\ClassConst\RemoveUnusedPrivateClassConstantRector;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
@@ -12,5 +12,5 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     $parameters->set(Option::PROJECT_TYPE, ProjectType::OPEN_SOURCE);
 
     $services = $containerConfigurator->services();
-    $services->set(RemoveUnusedClassConstantRector::class);
+    $services->set(RemoveUnusedPrivateClassConstantRector::class);
 };
